- moved methods from trait do parent class - made removeObjectFromSession() available again + added bc layer + marked as deprecated - marked getObjectFromSession() as deprecated and replaced it with the successor
