+++ bundles/CoreBundle/EventListener/Frontend/ElementListener.php
@@ -283,7 +283,7 @@ class ElementListener implements EventSubscriberInterface, LoggerAwareInterface
     {
         // object preview
         if ($objectId = $request->get('pimcore_object_preview')) {
-            if ($object = Service::getObjectFromSession($objectId)) {
+            if ($object = Service::getElementFromSession('object', $objectId)) {
 
                 $this->logger->debug('Loading object {object} ({objectId}) from session', [
                     'object' => $object->getFullPath(),
