+++ doc/Development_Documentation/05_Objects/01_Object_Classes/03_Layout_Elements/02_Preview_Iframe.md
@@ -32,7 +32,7 @@ class IFrameController extends \Pimcore\Controller\FrontendController
         $language = $context["language"];
 
         // get the current editing data, not the saved one! 
-        $object = Service::getObjectFromSession($objectId);
+        $object = Service::getElementFromSession('object', $objectId);
 
         $response =  '<h1>Title for language "' . $language . '": '  . $object->getTitle($language) . "</h1>";
