+++ doc/Development_Documentation/05_Objects/01_Object_Classes/05_Class_Settings/05_Preview.md
@@ -35,5 +35,5 @@ The above configuration (`/shop/product~p%o_id`)  will result in the URL `/shop/
 You can get the editing data (not the saved one) by calling
 
 ```php
-$object = Service::getObjectFromSession($objectId);
+$object = Service::getElementFromSession('object', $objectId);
 ```
